+++ packages/3d-map/index.js
@@ -1,3 +1,0 @@
-export * from "canvas-globe";
-export { default } from "canvas-globe";
-