+++ packages/3d-map/react.js
@@ -1,5 +1,0 @@
-"use client";
-
-export * from "canvas-globe/react";
-export { default } from "canvas-globe/react";
-