+++ packages/3d-maps/data/world.d.ts
@@ -1,0 +1,4 @@
+export * from "canvas-globe/data/world";
+export { default } from "canvas-globe/data/world";
+
+