+++ packages/3d-maps/index.js
@@ -1,0 +1,4 @@
+export * from "canvas-globe";
+export { default } from "canvas-globe";
+
+