+++ packages/3d-world-map/fx-interaction.js
@@ -1,0 +1,2 @@
+export * from "canvas-globe/fx/interaction";
+