+++ packages/3d-world-map/react.d.ts
@@ -1,0 +1,4 @@
+export * from "canvas-globe/react";
+export { default } from "canvas-globe/react";
+
+